Add comprehensive settings system and enhanced Toggl client functionality
New features:
- Settings API with full CRUD operations and validation
- Settings service with hierarchical configuration support
- Settings schema with category-based organization
- Database migration for settings table
- Enhanced Toggl client with improved error handling
- Frontend settings API service

Updates:
- Enhanced test routes for API validation
- Improved database models and schemas
- Updated sync service with better error handling
- Frontend dark mode improvements
- Docker configuration updates

Cleanup:
- Removed duplicate config files
- Organized client code structure

+++ database/init.sql
@@ -63,16 +63,16 @@ CREATE TABLE IF NOT EXISTS rates (
 -- Create time_entries cache table
 CREATE TABLE IF NOT EXISTS time_entries_cache (
     id SERIAL PRIMARY KEY,
-    toggl_id INTEGER UNIQUE NOT NULL,
+    toggl_id BIGINT UNIQUE NOT NULL,
     description TEXT,
     duration INTEGER NOT NULL, -- in seconds
     start_time TIMESTAMP NOT NULL,
     stop_time TIMESTAMP,
-    user_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
     user_name VARCHAR(255),
-    project_id INTEGER,
+    project_id BIGINT,
     project_name VARCHAR(255),
-    client_id INTEGER,
+    client_id BIGINT,
     client_name VARCHAR(255),
     workspace_id INTEGER NOT NULL,
     billable BOOLEAN DEFAULT FALSE,
